Split server/index.ts into createApp & httpServer parts
Needed for nuxt.js so it's possible to hook postgraphile module into
running express/connect instance that nuxt.js already created and
manages
